chore: add tracing and refactor state root
