+++ server/Cargo.toml
@@ -16,5 +16,8 @@ serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.148"
 sqlx = { version = "0.8.6", features = ["runtime-tokio", "postgres", "uuid", "chrono"]}
 tokio = { version = "1.48.0", features = ["full"] }
-tower-http = { version = "0.6.8", features = ["cors"] }
+tower = { version = "0.5.2", features = ["tracing"] }
+tower-http = { version = "0.6.8", features = ["cors", "trace"] }
+tracing = "0.1.44"
+tracing-subscriber = "0.3.22"
 uuid = { version = "1.19.0", features = ["serde", "v4"] }
